requirements.txt
streamlit
python-dotenv
langchain
pypdf
openai
chromadb
tiktoken